Create file to find size env var
