+++ include/minishell.h
@@ -6,7 +6,7 @@
 #include <readline/readline.h>
 #include <readline/history.h>
 
-
+int	find_size(char *str);
